Computer broke...so I'm pushing this commit somewhat earlier that anticipated
Edited the .keep files
Added a 404 page (I will patch .htaccess later to make this file work as intended)
Added plays as a new database table. It will add a new record every time a video is played, and allow you to reWatch scenes the you watch often
Removed a bunch of local dependencies from the project. Run ````npm install``` OR ```yarn``` to initialize all the required files. (Readme.md for more info)
Added https option in settings table, the option will now remove warnings in the console about using HTTPS and HTTP together. Only use this option if you have configured HTTPS on the server, because the setting only changes the url to start with HTTPS instead of HTTP. If HTTPS is not correctly configured, this WILL NOT fix that.
And so much more...